split core
