--- python/paddle/tensorrt/impls/core.py
@@ -1,570 +1,0 @@
-# Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import logging
-import os
-import sys
-
-import numpy as np
-
-current_dir = os.path.dirname(os.path.abspath(__file__))
-parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
-if parent_dir not in sys.path:
-    sys.path.append(parent_dir)
-
-import tensorrt as trt
-
-from paddle.base.log_helper import get_logger
-from paddle.tensorrt.register import converter_registry
-
-_logger = get_logger(
-    __name__, logging.INFO, fmt='%(asctime)s-%(levelname)s: %(message)s'
-)
-from paddle.tensorrt.converter_utils import (
-    append_ones,
-    broadcast,
-    get_axes_for_reduce_op,
-    get_dynamic_dims,
-    get_positive_dim,
-    get_trt_plugin,
-    has_dynamic_shape,
-)
-
-
-@converter_registry.register("pd_op.add", trt_version="8.x")
-@converter_registry.register("pd_op.add_", trt_version="8.x")
-def add_converter(network, paddle_op, inputs):
-    weight_shape = paddle_op.operands()[1].source().shape
-    input_shape = paddle_op.operands()[0].source().shape
-
-    weight_tensor = inputs[1]
-    input_tensor = inputs[0]
-    if type(inputs[1]) == trt.Weights:
-        weight_tensor = network.add_constant(
-            weight_shape, inputs[1]
-        ).get_output(0)
-    if type(inputs[0]) == trt.Weights:
-        input_tensor = network.add_constant(input_shape, inputs[0]).get_output(
-            0
-        )
-    lhs_val, rhs_val = broadcast(
-        network,
-        input_tensor,
-        weight_tensor,
-        input_tensor.name,
-        weight_tensor.name,
-    )
-
-    out = network.add_elementwise(
-        lhs_val, rhs_val, trt.ElementWiseOperation.SUM
-    )
-    return out
-
-
-@converter_registry.register("pd_op.relu", trt_version="8.x")
-def relu_converter(network, paddle_op, inputs):
-    out = network.add_activation(inputs[0], trt.ActivationType.RELU)
-    return out
-
-
-@converter_registry.register("pd_op.matmul", trt_version="8.x")
-def matmul_converter(network, paddle_op, inputs):
-    weight_shape = paddle_op.operands()[1].source().shape
-    transpose_x = paddle_op.attrs()["transpose_x"]
-    transpose_y = paddle_op.attrs()["transpose_y"]
-    self_matrix_op = (
-        trt.MatrixOperation.TRANSPOSE
-        if transpose_x
-        else trt.MatrixOperation.NONE
-    )
-    other_matrix_op = (
-        trt.MatrixOperation.TRANSPOSE
-        if transpose_y
-        else trt.MatrixOperation.NONE
-    )
-
-    weight_tensor = inputs[1]
-    if type(inputs[1]) == trt.Weights:
-        weight_tensor = network.add_constant(
-            weight_shape, inputs[1]
-        ).get_output(0)
-    lhs_val, rhs_val = broadcast(
-        network, inputs[0], weight_tensor, inputs[0].name, weight_tensor.name
-    )
-    out = network.add_matrix_multiply(
-        lhs_val, self_matrix_op, rhs_val, other_matrix_op
-    )
-    return out
-
-
-@converter_registry.register("pd_op.full_int_array", trt_version="8.x")
-def full_int_array_converter(network, paddle_op, inputs):
-    shape = paddle_op.attrs()["value"]
-    shape_weight = trt.Weights(np.array(shape, dtype=np.int32))
-    full_int_array_tensor = network.add_constant([len(shape)], shape_weight)
-    return full_int_array_tensor
-
-
-@converter_registry.register("pd_op.reshape", trt_version="8.x")
-def reshape_converter(network, paddle_op, inputs):
-    input_tensor, shape_tensor = inputs
-    input_shape = paddle_op.operands()[0].source().shape
-
-    output_shape = paddle_op.results()[0].shape
-    if network.has_implicit_batch_dimension:
-        output_shape = output_shape[1:]
-
-    if type(input_tensor) == trt.Weights:
-        input_tensor = network.add_constant(
-            input_shape, input_tensor
-        ).get_output(0)
-
-    shuffle_layer = network.add_shuffle(input_tensor)
-
-    try:
-        reshape_dims = (
-            paddle_op.operands()[1].source().get_defining_op().attrs()["value"]
-        )
-        shuffle_layer.reshape_dims = tuple(reshape_dims)
-    except Exception:
-        shuffle_layer.set_input(1, shape_tensor)
-
-    return shuffle_layer
-
-
-@converter_registry.register("pd_op.transpose", trt_version="8.x")
-def transpose_converter(network, paddle_op, inputs):
-    perm = paddle_op.attrs()["perm"]
-    transposed_tensor = network.add_shuffle(inputs[0])
-    transposed_tensor.second_transpose = perm
-    return transposed_tensor
-
-
-@converter_registry.register("pd_op.full", trt_version="8.x")
-def full_converter(network, paddle_op, inputs):
-    shape = paddle_op.attrs()["shape"]
-    value = paddle_op.attrs().get("value", 1.0)  # 默认值为1.0
-    full_tensor = network.add_constant(
-        shape, np.full(shape, value, dtype=np.float32)
-    )
-    return full_tensor
-
-
-@converter_registry.register("pd_op.scale", trt_version="8.x")
-def scale_converter(network, paddle_op, inputs):
-    scale = paddle_op.operands()[1].source().get_defining_op().attrs()["value"]
-    bias = paddle_op.attrs().get("bias", 0.0)
-    power = paddle_op.attrs().get("power", 1.0)
-
-    # Convert scale, bias, and power to TensorRT weights
-    scale_weight = trt.Weights(np.array([scale], dtype=np.float32))
-    bias_weight = trt.Weights(np.array([bias], dtype=np.float32))
-    power_weight = trt.Weights(np.array([power], dtype=np.float32))
-
-    scale_layer = network.add_scale(
-        inputs[0],
-        mode=trt.ScaleMode.UNIFORM,
-        shift=bias_weight,
-        scale=scale_weight,
-        power=power_weight,
-    )
-    return scale_layer
-
-
-@converter_registry.register("pd_op.softmax", trt_version="8.x")
-def softmax_converter(network, paddle_op, inputs):
-    axis = paddle_op.attrs().get("axis", 0)
-    if axis < 0:
-        axis = len(inputs[0].shape) + axis
-
-    softmax_layer = network.add_softmax(inputs[0])
-    softmax_layer.axes = 1 << axis
-    return softmax_layer
-
-
-@converter_registry.register(
-    "pd_op.layer_norm", trt_version="trt_version_ge=8.6"
-)
-def layernorm_converter(network, paddle_op, inputs):
-    input_a, scale, bias = inputs
-
-    begin_norm_axis = paddle_op.attrs().get("begin_norm_axis", 0)
-    epsilon = paddle_op.attrs().get("epsilon", 1e-5)
-    assert len(paddle_op.operands()) == 3
-    scale_shape = paddle_op.operands()[1].source().shape
-
-    scale_tensor = network.add_constant(scale_shape, scale).get_output(0)
-    bias_shape = paddle_op.operands()[2].source().shape
-    bias_tensor = network.add_constant(bias_shape, bias).get_output(0)
-
-    # dims = list(range( len(input_a.shape) - len(normalized_shape), len(input_a.shape)))
-    dims = list(range(len(input_a.shape)))[begin_norm_axis:]
-    axes = get_axes_for_reduce_op(dims)
-
-    scale_tensor = append_ones(
-        network,
-        scale_tensor,
-        f"{scale_tensor.name}_broadcast",
-        len(input_a.shape) - len(scale_tensor.shape),
-    )
-
-    bias_tensor = append_ones(
-        network,
-        bias_tensor,
-        f"{bias_tensor.name}_broadcast",
-        len(input_a.shape) - len(bias_tensor.shape),
-    )
-
-    layer_norm = network.add_normalization(
-        input_a, scale_tensor, bias_tensor, axes
-    )
-    layer_norm.epsilon = epsilon
-    layer_norm.compute_precision = trt.float32
-
-    return layer_norm
-
-
-@converter_registry.register("pd_op.conv2d", trt_version="8.x")
-def conv2d_converter(network, paddle_op, inputs):
-    input_tensor, weight = inputs
-    weight_shape = paddle_op.operands()[1].source().shape
-
-    padding = paddle_op.attrs().get("paddings", [0, 0])
-    stride = paddle_op.attrs().get("strides", [1, 1])
-    dilation = paddle_op.attrs().get("dilations", [1, 1])
-    groups = paddle_op.attrs().get("groups", 1)
-
-    # weight_tensor = network.add_constant(weight_shape, weight).get_output(0)
-    kernel_shape = trt.Dims((weight_shape[2], weight_shape[3]))
-
-    conv_layer = network.add_convolution_nd(
-        input_tensor, weight_shape[0], kernel_shape, weight
-    )
-    conv_layer.stride_nd = stride
-    conv_layer.padding_nd = padding
-    conv_layer.dilation_nd = dilation
-    conv_layer.num_groups = groups
-
-    return conv_layer
-
-
-@converter_registry.register("pd_op.nonzero", trt_version="8.x")
-def non_zero_converter(network, paddle_op, inputs):
-    input_tensor = inputs[0]
-    cast_layer = network.add_cast(input_tensor, trt.float32)
-    non_zero_layer = network.add_non_zero(cast_layer.get_output(0))
-
-    return non_zero_layer
-
-
-@converter_registry.register("pd_op.gather_nd", trt_version="8.x")
-def gather_nd_converter(network, paddle_op, inputs):
-    input_tensor, indices_tensor = inputs
-    shuffle_layer = network.add_shuffle(indices_tensor)
-    shuffle_layer.first_transpose = trt.Permutation([1, 0])
-    # import pdb;pdb.set_trace()
-    non_zero_layer = network.add_gather_v2(
-        input_tensor, shuffle_layer.get_output(0), trt.GatherMode.ND
-    )
-    return non_zero_layer
-
-
-@converter_registry.register("pd_op.pool2d", trt_version="8.x")
-def pool2d_converter(network, paddle_op, inputs):
-    input_tensor = inputs[0]
-    pooling_type = paddle_op.attrs().get("pooling_type", "max")
-    padding = paddle_op.attrs().get("paddings", [0, 0])
-    stride = paddle_op.attrs().get("strides", [1, 1])
-    ceil_mode = paddle_op.attrs().get("ceil_mode", False)
-    exclusive = paddle_op.attrs().get("exclusive")
-    adaptive = paddle_op.attrs().get("adaptive")
-    padding_algorithm = paddle_op.attrs().get("padding_algorithm")
-
-    input_shape = input_tensor.shape
-
-    # TODO attention for these codes
-    if not paddle_op.attrs().get("kernel_size") and len(inputs) == 2:
-        # the size of pool2d inputs is 2, means kernel size is the second input.
-        # kernel_size_tensor = inputs[1]
-        full_int_op = paddle_op.operands()[1].source().get_defining_op()
-        if full_int_op.name() == "pd_op.full_int_array":
-            kernel_size = full_int_op.attrs().get("value")
-        else:
-            raise Exception(
-                "the defining op of kernel size must be pd_op.full_int_array"
-            )
-    else:
-        kernel_size = paddle_op.attrs().get("kernel_size")
-
-    if len(stride) == 0 or stride[0] is None:
-        stride = kernel_size
-
-    if pooling_type == "max":
-        pooling_type = trt.PoolingType.MAX
-    elif pooling_type == "avg":
-        pooling_type = trt.PoolingType.AVERAGE
-    else:
-        raise ValueError(f"Unsupported pooling type: {pooling_type}")
-
-    if padding_algorithm == "VALID":
-        padding = [0, 0]
-
-    if adaptive:
-        output_size = kernel_size
-        stride = tuple(input_shape[-2 + i] // output_size[i] for i in range(2))
-        kernel_size = tuple(
-            input_shape[-2 + i] - (output_size[i] - 1) * stride[i]
-            for i in range(2)
-        )
-
-        pool_layer = network.add_pooling_nd(
-            input_tensor, pooling_type, window_size=kernel_size
-        )
-        pool_layer.stride_nd = stride
-        if pooling_type == "max":
-            pool_layer.padding_nd = padding
-    else:
-        pool_layer = network.add_pooling(
-            input_tensor, pooling_type, window_size=kernel_size
-        )
-        pool_layer.stride = stride
-        pool_layer.padding = padding
-        if exclusive:
-            pool_layer.average_count_excludes_padding = True
-        else:
-            pool_layer.average_count_excludes_padding = False
-        if ceil_mode:
-            pool_layer.padding_mode = trt.PaddingMode.EXPLICIT_ROUND_UP
-
-    return pool_layer
-
-
-@converter_registry.register("pd_op.batch_norm", trt_version="8.x")
-@converter_registry.register("pd_op.batch_norm_", trt_version="8.x")
-def batch_norm_converter(network, paddle_op, inputs):
-    input_tensor, mean, variance, scale, bias = inputs
-
-    scale_shape = paddle_op.operands()[3].source().shape
-
-    power = np.ones(scale_shape, dtype='float32')
-    power = trt.Weights(power)
-    input_tensor_shape = paddle_op.operands()[0].source().shape
-    if has_dynamic_shape(input_tensor_shape):
-        assert (
-            input_tensor.shape[1] != -1
-        ), "Channel dim can't be dynamic for batch norm."
-    # For BatchNorm1d ,reshape 1d to 2d
-    output_shape = input_tensor_shape
-
-    if not network.has_implicit_batch_dimension and len(input_tensor_shape) < 4:
-        assert (
-            len(get_dynamic_dims(input_tensor.shape)) <= 1
-        ), "BatchNorm1D with more than one dynamic dims is not currently supported."
-        reshape_layer = network.add_shuffle(input_tensor)
-        if len(input_tensor_shape) == 2:
-            reshape_layer.reshape_dims = (
-                input_tensor_shape[0],
-                input_tensor_shape[1],
-                1,
-                1,
-            )
-        else:  # len(input_tensor_shape) ==3
-            reshape_layer.reshape_dims = (
-                input_tensor_shape[0],
-                input_tensor_shape[1],
-                input_tensor_shape[2],
-                1,
-            )
-        input_tensor = reshape_layer.get_output(0)
-    # (self: tensorrt.tensorrt.INetworkDefinition, input: tensorrt.tensorrt.ITensor, mode: tensorrt.tensorrt.ScaleMode, shift: tensorrt.tensorrt.Weights = None, scale: tensorrt.tensorrt.Weights = None, power: tensorrt.tensorrt.Weights = None) -> tensorrt.tensorrt.IScaleLayer
-    batch_norm_layer = network.add_scale(
-        input_tensor, trt.ScaleMode.CHANNEL, bias, scale, power
-    )
-    # For BatchNorm1d,reshape output back to 1d
-    if not network.has_implicit_batch_dimension and len(output_shape) < 4:
-        reshape_output_layer = network.add_shuffle(
-            batch_norm_layer.get_output(0)
-        )
-        reshape_output_layer.reshape_dims = tuple(output_shape)
-        batch_norm_layer = reshape_output_layer
-
-    return batch_norm_layer
-
-
-@converter_registry.register("pd_op.flatten", trt_version="8.x")
-def flatten_converter(network, paddle_op, inputs):
-    input_val = inputs[0]
-    input_val_shape = input_val.shape
-    dims = len(input_val_shape)
-
-    start_axis = paddle_op.attrs().get("start_axis")
-    stop_axis = paddle_op.attrs().get("stop_axis")
-
-    flatten_layer = network.add_shuffle(input_val)
-
-    if not has_dynamic_shape(input_val_shape):
-        if start_axis < 0:
-            start_axis += dims + 1
-        if stop_axis < 0:
-            stop_axis += dims + 1
-
-        flatten_dim = 1
-        final_shape = []
-
-        for i, s in enumerate(input_val_shape):
-            if i >= start_axis and i <= stop_axis:
-                flatten_dim *= s
-            elif i == stop_axis + 1:
-                final_shape.append(flatten_dim)
-                final_shape.append(s)
-            else:
-                final_shape.append(s)
-
-        if stop_axis == len(input_val.shape) - 1:
-            final_shape.append(flatten_dim)
-
-        flatten_layer.reshape_dims = tuple(final_shape)
-    else:
-        input_shape_layer = network.add_shape(input_val)
-        input_shape_layer.name = f"{input_val.name}_origin_shape"
-
-        final_shapes = []
-        # Shapes before start_axis
-        if start_axis > 0:
-            prefix_shape_layer = network.add_slice(
-                input_shape_layer.get_output(0),
-                start=(0,),
-                shape=(start_axis,),
-                stride=(1,),
-            )
-            prefix_shape_layer.name = f"{input_val.name}_prefix_shape"
-            final_shapes.append(prefix_shape_layer.get_output(0))
-
-        flatten_shape_layer = network.add_slice(
-            input_shape_layer.get_output(0),
-            start=(start_axis,),
-            shape=(stop_axis - start_axis + 1,),
-            stride=(1,),
-        )
-        flatten_shape_layer.name = f"{input_val.name}_need_flatten"
-        flatten_shape_layer = network.add_reduce(
-            flatten_shape_layer.get_output(0),
-            trt.ReduceOperation.PROD,
-            axes=get_axes_for_reduce_op(0, False),
-            keep_dims=True,
-        )
-        flatten_shape_layer.name = f"{input_val.name}_flatten_dim"
-        final_shapes.append(flatten_shape_layer.get_output(0))
-
-        # Shapes after stop_axis
-        if stop_axis < len(input_val_shape) - 1:
-            suffix_shape_layer = network.add_slice(
-                input_shape_layer.get_output(0),
-                start=(stop_axis + 1,),
-                shape=(len(input_val_shape) - stop_axis - 1,),
-                stride=(1,),
-            )
-            suffix_shape_layer.name = f"{input_val.name}_suffix_shape"
-            final_shapes.append(suffix_shape_layer.get_output(0))
-
-        final_shape_layer = network.add_concatenation(final_shapes)
-        final_shape_layer.axis = 0
-        final_shape_layer.name = f"{input_val.name}_final_shape"
-        flatten_layer.set_input(1, final_shape_layer.get_output(0))
-
-    return flatten_layer
-
-
-# In the converter, pd_op.concat has three inputs, because builtin.combine has two inputs.
-@converter_registry.register("pd_op.concat", trt_version="8.x")
-def concat_converter(network, paddle_op, inputs):
-    input_tensors = inputs[:-1]
-    axis_tensor = inputs[-1]
-    concat_layer = network.add_concatenation(inputs=input_tensors)
-
-    axis = paddle_op.operands()[1].source().get_defining_op().attrs()["value"]
-    axis = int(axis)
-    if axis < 0:
-        axis = len(input_tensors[0].shape) + axis
-    concat_layer.axis = axis
-
-    return concat_layer
-
-
-@converter_registry.register("pd_op.gelu", trt_version="8.x")
-def gelu_converter(network, paddle_op, inputs):
-    input_val = inputs[0]
-    approximate = paddle_op.attrs()["approximate"]
-    if approximate:
-        raise RuntimeError(
-            "GeLU converter currently doesn't support fast gelu compute"
-        )
-
-    plugin_name = "CustomGeluPluginDynamic"
-    type_id = trt.PluginField(
-        "type_id", np.array(0, dtype=np.int32), trt.PluginFieldType.INT32
-    )
-
-    filed_collection = trt.PluginFieldCollection([type_id])
-    plugin_version = "1"
-
-    plugin = get_trt_plugin(plugin_name, filed_collection, plugin_version)
-
-    layer = network.add_plugin_v2([input_val], plugin)
-    return layer
-
-
-@converter_registry.register("pd_op.unsqueeze", trt_version="8.x")
-@converter_registry.register("pd_op.unsqueeze_", trt_version="8.x")
-def unsqueeze_converter(network, paddle_op, inputs):
-    input_val = inputs[0]
-    input_shape = paddle_op.operands()[0].source().shape
-    input_shape_size = len(input_shape)
-
-    if type(input_val) == trt.Weights:
-        input_val = network.add_constant(input_shape, input_val).get_output(0)
-    axis = paddle_op.operands()[1].source().get_defining_op().attrs()["value"]
-    axis = axis[0]
-
-    axis = get_positive_dim(axis, input_shape_size + 1)
-    layer = network.add_shuffle(input_val)
-    layer.reshape_dims = (
-        tuple(input_val.shape)[:axis] + (1,) + tuple(input_val.shape)[axis:]
-    )
-    return layer
-
-
-@converter_registry.register("pd_op.squeeze", trt_version="8.x")
-@converter_registry.register("pd_op.squeeze_", trt_version="8.x")
-def squeeze_converter(network, paddle_op, inputs):
-    input_val = inputs[0]
-    input_shape = paddle_op.operands()[0].source().shape
-    input_shape_size = len(input_shape)
-
-    if type(input_val) == trt.Weights:
-        input_val = network.add_constant(input_shape, input_val).get_output(0)
-
-    axis = paddle_op.operands()[1].source().get_defining_op().attrs()["value"]
-    axis = axis[0]
-
-    axis = get_positive_dim(axis, input_shape_size + 1)
-    output_shape = []
-    for i, s in enumerate(input_shape):
-        if i == axis and s == 1:
-            continue
-        output_shape.append(s)
-
-    layer = network.add_shuffle(input_val)
-    layer.reshape_dims = tuple(output_shape)
-    return layer
