+++ python/paddle/tensorrt/impls/creation.py
@@ -1,0 +1,50 @@
+# Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+import os
+import sys
+
+import numpy as np
+
+current_dir = os.path.dirname(os.path.abspath(__file__))
+parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
+if parent_dir not in sys.path:
+    sys.path.append(parent_dir)
+
+import tensorrt as trt
+
+from paddle.base.log_helper import get_logger
+from paddle.tensorrt.register import converter_registry
+
+_logger = get_logger(
+    __name__, logging.INFO, fmt='%(asctime)s-%(levelname)s: %(message)s'
+)
+
+
+@converter_registry.register("pd_op.full_int_array", trt_version="8.x")
+def full_int_array_converter(network, paddle_op, inputs):
+    shape = paddle_op.attrs()["value"]
+    shape_weight = trt.Weights(np.array(shape, dtype=np.int32))
+    full_int_array_tensor = network.add_constant([len(shape)], shape_weight)
+    return full_int_array_tensor
+
+
+@converter_registry.register("pd_op.full", trt_version="8.x")
+def full_converter(network, paddle_op, inputs):
+    shape = paddle_op.attrs()["shape"]
+    value = paddle_op.attrs().get("value", 1.0)  # 默认值为1.0
+    full_tensor = network.add_constant(
+        shape, np.full(shape, value, dtype=np.float32)
+    )
+    return full_tensor
